#7 Address feedback.

+++ run.sh
@@ -1,18 +1,41 @@
 #!/usr/bin/env bash
 
+# Exit on Error.
+set -e
+
+# Read our .env file.
+export $(grep -v '^#' .env | xargs)
+
+# Build the Docker images with the current user's ID.
 docker compose build --build-arg USER_ID=$(id -u ${USER})
 
+# Start the containers in detached mode.
 docker compose up -d
 
-until $(docker exec -i vmrt-emr-process-log-mysql mysql -uroot -pbmorris -e "DROP DATABASE IF EXISTS vmrt_emr_transcription; CREATE DATABASE vmrt_emr_transcription;"); do
-  echo 'Waiting for database container to start...'
+# Wait for the database container to start (with a timeout).
+TIMEOUT=30
+COUNTER=0
+until $(docker exec -i vmrt-emr-process-log-mysql mysql -uroot -p$SQL_PASSWORD -e "DROP DATABASE IF EXISTS vmrt_emr_transcription; CREATE DATABASE vmrt_emr_transcription;") || [[ $COUNTER -eq $TIMEOUT ]]; do
+  echo "Waiting for database container to start... ($COUNTER/$TIMEOUT)"
   sleep 1
+  COUNTER=$((COUNTER+1))
 done
 
-echo 'Done!'
+if [[ $COUNTER -eq $TIMEOUT ]]; then
+  echo "Error: Timeout waiting for database container."
+  exit 1
+fi
 
-docker exec -t vmrt-emr-workspace python /workspace/update.py
+echo "Database initialized successfully."
 
+# Execute the Python script.
+if ! docker exec -t vmrt-emr-workspace python ./scripts/database_setup.py install; then
+  echo "Error: Failed to execute Python script."
+  exit 1
+fi
+
+# Provide an interactive Bash shell within the container.
 docker exec -it vmrt-emr-workspace bash
 
+# Stop and remove the containers.
 docker compose down
